requirements.txt
asgiref==3.5.2
black==22.8.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1.1
django-debug-toolbar==3.7.0
django-debug-toolbar-template-profiler==2.1.0
django-filter==22.1
django-widget-tweaks==1.4.12
idna==3.4
mypy-extensions==0.4.3
oauthlib==3.2.1
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.5.0
python-decouple==3.6
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.2
urllib3==1.26.12
wrapt==1.14.1
